[exchange] use BigDecimal as application numeric type

+++ src/main/java/com/crypto/currency/dto/CurrencyRatesResponse.java
@@ -1,11 +1,13 @@
 package com.crypto.currency.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 public class CurrencyRatesResponse {
     private String source;
-    private Map<String, Double> rates;
+    private Map<String, BigDecimal> rates;
 }
