+++ src/main/java/com/crypto/currency/provider/coingecko/CoinGeckoProvider.java
@@ -5,6 +5,7 @@ import com.crypto.currency.provider.coingecko.client.CoinGeckoApiClient;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
@@ -17,13 +18,13 @@ public class CoinGeckoProvider implements CryptoProvider {
     private static final String DEFAULT_RATE = "usd";
 
     @Override
-    public Map<String, Double> getRatesForCurrency(String baseSymbol, List<String> rates) {
+    public Map<String, BigDecimal> getRatesForCurrency(String baseSymbol, List<String> rates) {
         String coinGeckoId = mapToCoinGeckoId(baseSymbol);
 
         if (rates == null || rates.isEmpty()) {
             rates = List.of(DEFAULT_RATE);
         }
-        Map<String, Map<String, Double>> response = apiClient.getSimplePrice(coinGeckoId, rates);
+        Map<String, Map<String, BigDecimal>> response = apiClient.getSimplePrice(coinGeckoId, rates);
 
         if (response == null || !response.containsKey(coinGeckoId)) {
             return Collections.emptyMap();
